fix lints

+++ src/file.rs
@@ -55,6 +55,6 @@ impl FileState {
     /// gets the number of digits used to represent the file addresses
     /// (rounded up to be in pairs
     pub fn address_digits(&self) -> u8 {
-        (ilog2((self.content.len() - 1).max(1)) / 8 + 1) as u8 * 2
+        (ilog2((self.content.len() - 1).max(1)) / 8 + 1) * 2
     }
 }
